adding particles as gaze indicator
Summary:
adding a particle system that indicates where the gaze / ray casting is happening. This would help the game to be more intuitive by providing a continues visual feedback to the user to prevent issues reported in T147032423.

* Worked on the following UX suggestions by aimanb
Inaccurate object selection: Gaze indicator and hands are too far from object yet it is marked as selected
(Sorob): this was the result of object locking. A feature that kept the selected object in focus to finish recording audio even after the gaze was no longer on it. I removed this.
* Parity in Hand activation and particle visibility: Particle is active while hands are inactive
(Sorob): This was caused by long particle animations. I shortened it. Now you only see the particles fading out for a second after you stop the gesture
* Lag in where the hands position + gaze compared to where particles are positioned and when object is selected
Alt: Reduce this latency to give feeling of more control
(Sorob): Removed latency
* Only use particle effect to map gaze movement and disappear once object is selected
Sorob: DONE

Reviewed By: yolanother

Differential Revision: D44427482

fbshipit-source-id: 574de4c060c4c6af3f4678a2d2c74ffb6c221dca